speech_to_text/requirements.txt
google-cloud-speech
flask